refactor(datastorage): split config.go into secrets/validate/accessors (Wave 3)
Split pkg/datastorage/config/config.go (814 lines) into config.go (trimmed:
struct family + LoadFromFile), config_secrets.go (LoadSecrets + per-source
loaders), config_validate.go (Validate + per-section validators), and
config_accessors.go (Get* derived-value accessors).

Pure code motion, no behavior change, no exported-API change. See audit
report GO-ANTIPATTERN-AUDIT-2026-07-01 §7k-3.
