+++ pkg/datastorage/config/config.go
@@ -19,7 +19,6 @@ package config
 import (
 	"crypto/tls"
 	"crypto/x509"
-	"encoding/json"
 	"fmt"
 	"os"
 	"strings"
@@ -283,509 +282,7 @@ func LoadFromFile(path string) (*Config, error) {
 	return &cfg, nil
 }
 
-// LoadSecrets loads secrets from mounted Kubernetes Secret files (ADR-030 Section 6)
-// It supports both YAML and JSON secret files.
-// This function REQUIRES secretsFile to be configured for both database and redis.
-func (c *Config) LoadSecrets() error {
-	// Load database secrets (REQUIRED)
-	if c.Database.SecretsFile == "" {
-		return fmt.Errorf("database secretsFile required (ADR-030 Section 6)")
-	}
-	if c.Database.PasswordKey == "" {
-		return fmt.Errorf("database passwordKey required (ADR-030 Section 6)")
-	}
-
-	dbSecrets, err := loadSecretFile(c.Database.SecretsFile)
-	if err != nil {
-		return fmt.Errorf("failed to load database secrets from %s: %w",
-			c.Database.SecretsFile, err)
-	}
-
-	// Extract password using configured key
-	password, ok := dbSecrets[c.Database.PasswordKey]
-	if !ok {
-		return fmt.Errorf("password key '%s' not found in database secret file %s",
-			c.Database.PasswordKey, c.Database.SecretsFile)
-	}
-
-	passwordStr, isString := password.(string)
-	if !isString {
-		return fmt.Errorf("database password key '%s' in secret file is not a string",
-			c.Database.PasswordKey)
-	}
-
-	c.Database.Password = passwordStr
-
-	// Optional: Override username from secret if key specified
-	if c.Database.UsernameKey != "" {
-		if username, ok := dbSecrets[c.Database.UsernameKey]; ok {
-			if usernameStr, usernameIsString := username.(string); usernameIsString {
-				c.Database.User = usernameStr
-			} else {
-				return fmt.Errorf("database username key '%s' in secret file is not a string",
-					c.Database.UsernameKey)
-			}
-		}
-	}
-
-	// Load Redis secrets (REQUIRED)
-	if c.Redis.SecretsFile == "" {
-		return fmt.Errorf("redis secretsFile required (ADR-030 Section 6)")
-	}
-	if c.Redis.PasswordKey == "" {
-		return fmt.Errorf("redis passwordKey required (ADR-030 Section 6)")
-	}
-
-	redisSecrets, err := loadSecretFile(c.Redis.SecretsFile)
-	if err != nil {
-		return fmt.Errorf("failed to load redis secrets from %s: %w",
-			c.Redis.SecretsFile, err)
-	}
-
-	// Extract Redis password
-	redisPassword, ok := redisSecrets[c.Redis.PasswordKey]
-	if !ok {
-		return fmt.Errorf("password key '%s' not found in redis secret file %s",
-			c.Redis.PasswordKey, c.Redis.SecretsFile)
-	}
-
-	redisPasswordStr, isString := redisPassword.(string)
-	if !isString {
-		return fmt.Errorf("redis password key '%s' in secret file is not a string",
-			c.Redis.PasswordKey)
-	}
-
-	c.Redis.Password = redisPasswordStr
-
-	// GAP-05 (Issue #1505): Load optional HMAC key for keyed audit hash chain.
-	// Unlike database/redis secrets, this is OPTIONAL — omitting it preserves
-	// the legacy unkeyed SHA256 algorithm for backward compatibility.
-	if c.Audit.HashKeySecretsFile != "" {
-		if c.Audit.HashKeyKey == "" {
-			return fmt.Errorf("audit hashKeyKey required when hashKeySecretsFile is set (GAP-05)")
-		}
-
-		auditSecrets, err := loadSecretFile(c.Audit.HashKeySecretsFile)
-		if err != nil {
-			return fmt.Errorf("failed to load audit hash key secrets from %s: %w",
-				c.Audit.HashKeySecretsFile, err)
-		}
-
-		hmacKeyRaw, ok := auditSecrets[c.Audit.HashKeyKey]
-		if !ok {
-			return fmt.Errorf("hash key '%s' not found in audit secret file %s",
-				c.Audit.HashKeyKey, c.Audit.HashKeySecretsFile)
-		}
-
-		hmacKeyStr, isString := hmacKeyRaw.(string)
-		if !isString {
-			return fmt.Errorf("audit hash key '%s' in secret file is not a string",
-				c.Audit.HashKeyKey)
-		}
-		if hmacKeyStr == "" {
-			return fmt.Errorf("audit hash key '%s' in secret file %s is empty",
-				c.Audit.HashKeyKey, c.Audit.HashKeySecretsFile)
-		}
-
-		c.Audit.HMACKey = []byte(hmacKeyStr)
-	}
-
-	return nil
-}
-
-// loadSecretFile unmarshals a secret file (supports YAML and JSON)
-// This is a helper function for LoadSecrets()
-func loadSecretFile(secretFilePath string) (map[string]interface{}, error) {
-	data, err := os.ReadFile(secretFilePath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read secret file %s: %w", secretFilePath, err)
-	}
-
-	var secrets map[string]interface{}
-
-	// Try YAML first
-	if err := yaml.Unmarshal(data, &secrets); err == nil {
-		return secrets, nil
-	}
-
-	// Fallback to JSON
-	if err := json.Unmarshal(data, &secrets); err != nil {
-		return nil, fmt.Errorf("failed to parse secret file as YAML or JSON: %w", err)
-	}
-
-	return secrets, nil
-}
-
-// Validate checks if the configuration is valid and returns an error if not
-// ADR-030: Validate configuration before service startup
-//
-// Decomposed into per-section validators (GO-ANTIPATTERN-AUDIT-2026-07-01
-// Phase 4a) purely for readability/complexity — behavior, order, and error
-// messages are unchanged from the pre-decomposition implementation.
-func (c *Config) Validate() error {
-	if err := c.validateDatabase(); err != nil {
-		return err
-	}
-	if err := c.validateServer(); err != nil {
-		return err
-	}
-	if err := c.validateRedis(); err != nil {
-		return err
-	}
-	if err := c.validateProductionConstraints(); err != nil {
-		return err
-	}
-	if err := c.validateRetention(); err != nil {
-		return err
-	}
-	if err := c.validateLogging(); err != nil {
-		return err
-	}
-	if err := c.validateTimeouts(); err != nil {
-		return err
-	}
-	return c.validateCORS()
-}
-
-// validateDatabase checks required PostgreSQL connection fields and applies
-// safe defaults for the connection pool and per-statement/per-lock timeouts.
-func (c *Config) validateDatabase() error {
-	if c.Database.Host == "" {
-		return fmt.Errorf("database host required")
-	}
-	if c.Database.Port == 0 {
-		return fmt.Errorf("database port required")
-	}
-	if c.Database.Port < 1 || c.Database.Port > 65535 {
-		return fmt.Errorf("database port must be between 1 and 65535, got: %d", c.Database.Port)
-	}
-	if c.Database.Name == "" {
-		return fmt.Errorf("database name required")
-	}
-	if c.Database.User == "" {
-		return fmt.Errorf("database user required")
-	}
-
-	// Issue #667/M3: Default MaxOpenConns to 25 to prevent unlimited connections
-	if c.Database.MaxOpenConns == 0 {
-		c.Database.MaxOpenConns = 25
-	}
-	if c.Database.MaxOpenConns < 0 {
-		return fmt.Errorf("database maxOpenConns must be positive, got: %d", c.Database.MaxOpenConns)
-	}
-
-	// Issue #667/M1: Default statement_timeout and lock_timeout for PG-level safety
-	if c.Database.StatementTimeout == "" {
-		c.Database.StatementTimeout = "30s"
-	}
-	if _, err := time.ParseDuration(c.Database.StatementTimeout); err != nil {
-		return fmt.Errorf("invalid statementTimeout: %w", err)
-	}
-	if c.Database.LockTimeout == "" {
-		c.Database.LockTimeout = "10s"
-	}
-	if _, err := time.ParseDuration(c.Database.LockTimeout); err != nil {
-		return fmt.Errorf("invalid lockTimeout: %w", err)
-	}
-	return nil
-}
-
-// validateServer checks required HTTP server fields and applies safe
-// defaults for the metrics/health ports and max batch size.
-func (c *Config) validateServer() error {
-	if c.Server.Port == 0 {
-		return fmt.Errorf("server port required")
-	}
-	if c.Server.Port < 1024 || c.Server.Port > 65535 {
-		return fmt.Errorf("server port must be between 1024 and 65535, got: %d", c.Server.Port)
-	}
-
-	// Issue #283: Default metricsPort to 9090 (Kubernaut standard) when omitted
-	if c.Server.MetricsPort == 0 {
-		c.Server.MetricsPort = 9090
-	}
-
-	// Issue #753: Default healthPort to 8081 (CONFIG_STANDARDS.md) when omitted
-	if c.Server.HealthPort == 0 {
-		c.Server.HealthPort = 8081
-	}
-
-	// Issue #667 / BR-STORAGE-043: Default MaxBatchSize to 500 when omitted
-	if c.Server.MaxBatchSize == 0 {
-		c.Server.MaxBatchSize = 500
-	}
-	if c.Server.MaxBatchSize < 0 {
-		return fmt.Errorf("server maxBatchSize must be positive, got: %d", c.Server.MaxBatchSize)
-	}
-	if c.Server.MetricsPort < 1024 || c.Server.MetricsPort > 65535 {
-		return fmt.Errorf("server metricsPort must be between 1024 and 65535, got: %d", c.Server.MetricsPort)
-	}
-	return nil
-}
-
-// validateRedis checks required Redis fields and rejects a half-configured
-// TLS setup (enabled but no trust anchor) regardless of environment (SC-8).
-func (c *Config) validateRedis() error {
-	if c.Redis.Addr == "" {
-		return fmt.Errorf("redis address required")
-	}
-	if c.Redis.TLS.Enabled {
-		if c.Redis.TLS.CAFile == "" {
-			return fmt.Errorf("redis TLS enabled but no caFile specified; mount the CA certificate (SC-8)")
-		}
-	}
-	return nil
-}
-
-// validateProductionConstraints rejects insecure transport/CORS
-// configurations when Environment is "production" (FED-C1/SC-8, AC-4).
-func (c *Config) validateProductionConstraints() error {
-	if !c.IsProduction() {
-		return nil
-	}
-	if c.Database.SSLMode == "" || c.Database.SSLMode == "disable" {
-		return fmt.Errorf("database sslMode must not be 'disable' in production (SC-8); use verify-full or verify-ca")
-	}
-	if !c.Redis.TLS.Enabled {
-		return fmt.Errorf("redis TLS must be enabled in production (SC-8); set redis.tls.enabled=true")
-	}
-	for _, origin := range c.Server.CORSAllowedOrigins {
-		if origin == "*" {
-			return fmt.Errorf("CORS wildcard origin '*' is not allowed in production (AC-4); use explicit origins")
-		}
-	}
-	return nil
-}
-
-// validateRetention checks the audit-retention worker configuration (AU-11).
-func (c *Config) validateRetention() error {
-	if c.Retention.Interval != "" {
-		if _, err := time.ParseDuration(c.Retention.Interval); err != nil {
-			return fmt.Errorf("invalid retention interval: %w", err)
-		}
-	}
-	if c.Retention.DefaultDays < 0 {
-		return fmt.Errorf("retention defaultDays must be non-negative, got: %d", c.Retention.DefaultDays)
-	}
-	return nil
-}
-
-// validateLogging checks the configured log level (case-insensitive per Issue #875).
-func (c *Config) validateLogging() error {
-	validLevels := map[string]bool{
-		"DEBUG": true,
-		"INFO":  true,
-		"WARN":  true,
-		"ERROR": true,
-	}
-	if c.Logging.Level != "" && !validLevels[strings.ToUpper(c.Logging.Level)] {
-		return fmt.Errorf("invalid log level: %s (must be DEBUG, INFO, WARN, or ERROR)", c.Logging.Level)
-	}
-	return nil
-}
-
-// validateTimeouts parses every configurable duration/size field to fail
-// fast at startup instead of silently defaulting at runtime (SI-10).
-func (c *Config) validateTimeouts() error {
-	if c.Server.ReadTimeout != "" {
-		if _, err := time.ParseDuration(c.Server.ReadTimeout); err != nil {
-			return fmt.Errorf("invalid readTimeout: %w", err)
-		}
-	}
-	if c.Server.WriteTimeout != "" {
-		if _, err := time.ParseDuration(c.Server.WriteTimeout); err != nil {
-			return fmt.Errorf("invalid writeTimeout: %w", err)
-		}
-	}
-	if c.Database.ConnMaxLifetime != "" {
-		if _, err := time.ParseDuration(c.Database.ConnMaxLifetime); err != nil {
-			return fmt.Errorf("invalid connMaxLifetime: %w", err)
-		}
-	}
-	if c.Database.ConnMaxIdleTime != "" {
-		if _, err := time.ParseDuration(c.Database.ConnMaxIdleTime); err != nil {
-			return fmt.Errorf("invalid connMaxIdleTime: %w", err)
-		}
-	}
-
-	// SI-10: Fail fast on invalid shutdown/propagation durations instead of
-	// silently defaulting at runtime where operators may not notice.
-	if c.Server.ShutdownTimeout != "" {
-		if _, err := time.ParseDuration(c.Server.ShutdownTimeout); err != nil {
-			return fmt.Errorf("invalid shutdownTimeout: %w", err)
-		}
-	}
-	if c.Server.EndpointPropagationDelay != "" {
-		if _, err := time.ParseDuration(c.Server.EndpointPropagationDelay); err != nil {
-			return fmt.Errorf("invalid endpointPropagationDelay: %w", err)
-		}
-	}
-	if c.Server.MaxBodySize != "" {
-		var size int64
-		if n, err := fmt.Sscanf(c.Server.MaxBodySize, "%d", &size); err != nil || n != 1 {
-			return fmt.Errorf("invalid maxBodySize %q: must be an integer (bytes)", c.Server.MaxBodySize)
-		}
-	}
-	return nil
-}
-
-// validateCORS checks each configured CORS origin's format, aligned with
-// go-chi/cors semantics (AC-4/SI-10).
-func (c *Config) validateCORS() error {
-	for i, origin := range c.Server.CORSAllowedOrigins {
-		trimmed := strings.TrimSpace(origin)
-		if trimmed == "" {
-			return fmt.Errorf("corsAllowedOrigins[%d]: empty or whitespace-only origin", i)
-		}
-		if trimmed == "*" {
-			continue
-		}
-		lower := strings.ToLower(trimmed)
-		if !strings.HasPrefix(lower, "http://") && !strings.HasPrefix(lower, "https://") {
-			return fmt.Errorf("corsAllowedOrigins[%d] %q: must start with http:// or https:// (or be \"*\")", i, origin)
-		}
-	}
-	return nil
-}
-
-// GetReadTimeout returns the read timeout as a time.Duration
-func (c *ServerConfig) GetReadTimeout() time.Duration {
-	if c.ReadTimeout == "" {
-		return 30 * time.Second // default
-	}
-	duration, err := time.ParseDuration(c.ReadTimeout)
-	if err != nil {
-		return 30 * time.Second // fallback to default
-	}
-	return duration
-}
-
-// GetWriteTimeout returns the write timeout as a time.Duration
-func (c *ServerConfig) GetWriteTimeout() time.Duration {
-	if c.WriteTimeout == "" {
-		return 30 * time.Second // default
-	}
-	duration, err := time.ParseDuration(c.WriteTimeout)
-	if err != nil {
-		return 30 * time.Second // fallback to default
-	}
-	return duration
-}
-
-// GetShutdownTimeout returns the graceful shutdown budget as a time.Duration.
-// DD-007: Defaults to 60s. Clamped to [30s, 120s] for safety — a value below 30s
-// risks data loss (DLQ drain alone needs 10s), a value above 120s wastes K8s resources.
-func (c *ServerConfig) GetShutdownTimeout() time.Duration {
-	const (
-		defaultTimeout = 60 * time.Second
-		minTimeout     = 30 * time.Second
-		maxTimeout     = 120 * time.Second
-	)
-	if c.ShutdownTimeout == "" {
-		return defaultTimeout
-	}
-	duration, err := time.ParseDuration(c.ShutdownTimeout)
-	if err != nil {
-		return defaultTimeout
-	}
-	if duration < minTimeout {
-		return minTimeout
-	}
-	if duration > maxTimeout {
-		return maxTimeout
-	}
-	return duration
-}
-
-// GetEndpointPropagationDelay returns the endpoint propagation delay as time.Duration.
-// #1088 Phase 7 / SRE-L1: Defaults to 5s. Clamped to [1s, 30s] — below 1s risks
-// receiving traffic before K8s propagation; above 30s wastes shutdown budget.
-func (c *ServerConfig) GetEndpointPropagationDelay() time.Duration {
-	const (
-		defaultDelay = 5 * time.Second
-		minDelay     = 1 * time.Second
-		maxDelay     = 30 * time.Second
-	)
-	if c.EndpointPropagationDelay == "" {
-		return defaultDelay
-	}
-	duration, err := time.ParseDuration(c.EndpointPropagationDelay)
-	if err != nil {
-		return defaultDelay
-	}
-	if duration < minDelay {
-		return minDelay
-	}
-	if duration > maxDelay {
-		return maxDelay
-	}
-	return duration
-}
-
-// GetMaxBodySize returns the maximum request body size in bytes.
-// #1048 Phase 4 / SC-5: Defaults to 5 MiB. Clamped to [1 MiB, 50 MiB].
-// 5 MiB accommodates batch audit event requests (500 events × ~2-5 KB each).
-func (c *ServerConfig) GetMaxBodySize() int64 {
-	const (
-		mib         = 1 << 20
-		defaultSize = 5 * mib
-		minSize     = 1 * mib
-		maxSize     = 50 * mib
-	)
-	if c.MaxBodySize == "" {
-		return int64(defaultSize)
-	}
-	var size int64
-	if n, err := fmt.Sscanf(c.MaxBodySize, "%d", &size); err == nil && n == 1 {
-		// Plain integer in bytes
-	} else {
-		return int64(defaultSize)
-	}
-	if size < int64(minSize) {
-		return int64(minSize)
-	}
-	if size > int64(maxSize) {
-		return int64(maxSize)
-	}
-	return size
-}
-
-// GetCORSAllowedOrigins returns the configured CORS origins.
-// SEC-H3/AC-4: Defaults to empty (reject all cross-origin requests).
-// Operators must configure explicit origins for browser-based access.
-func (c *ServerConfig) GetCORSAllowedOrigins() []string {
-	return c.CORSAllowedOrigins
-}
-
-// GetSignerCertDir returns the directory holding tls.crt/tls.key for audit export signing.
-// #1048 Phase 5 / AU-9: Defaults to /etc/certs when unset.
-func (c *ServerConfig) GetSignerCertDir() string {
-	if c.SignerCertDir == "" {
-		return "/etc/certs"
-	}
-	return c.SignerCertDir
-}
-
-// GetConnectionString returns the PostgreSQL connection string with PG-level timeouts.
-// Issue #667/M1: statement_timeout and lock_timeout are set as DSN options so PostgreSQL
-// itself enforces limits, independent of Go-side context cancellation.
-func (c *DatabaseConfig) GetConnectionString() string {
-	dsn := fmt.Sprintf("host=%s port=%d dbname=%s user=%s password=%s sslmode=%s",
-		c.Host, c.Port, c.Name, c.User, c.Password, c.SSLMode)
-
-	var pgOpts []string
-	if c.StatementTimeout != "" {
-		if d, err := time.ParseDuration(c.StatementTimeout); err == nil {
-			pgOpts = append(pgOpts, fmt.Sprintf("-c statement_timeout=%d", d.Milliseconds()))
-		}
-	}
-	if c.LockTimeout != "" {
-		if d, err := time.ParseDuration(c.LockTimeout); err == nil {
-			pgOpts = append(pgOpts, fmt.Sprintf("-c lock_timeout=%d", d.Milliseconds()))
-		}
-	}
-	if len(pgOpts) > 0 {
-		dsn += fmt.Sprintf(" options='%s'", strings.Join(pgOpts, " "))
-	}
-	return dsn
-}
+// LoadSecrets, Validate, and the ServerConfig/DatabaseConfig accessor methods
+// are defined in config_secrets.go, config_validate.go, and
+// config_accessors.go respectively (GO-ANTIPATTERN-AUDIT-2026-07-01 Wave 3,
+// pure code motion, no behavior change).
